qdio: extend API to allow polling
Patchwork-id: 29149
O-Subject: [RHEL6.1 PATCH 06/10] [s390x] qdio: extend API to allow polling
Bugzilla: 633525

Description
-----------
Extend the qdio API to allow polling in the upper-layer driver.
This is needed by qeth to use NAPI.  To use the new interface
the upper-layer driver must specify the queue_start_poll().
This callback is used to signal the upper-layer driver that
it has initiative and must process the inbound queue by calling
qdio_get_next_buffers().  If the upper-layer driver wants to stop
polling it calls qdio_start_irq().

Because adapter interrupts are not completely stoppable qdio
implements a software bit QDIO_QUEUE_IRQS_DISABLED to safely
disable interrupts for an input queue.
The old interface is preserved and will be used as is by zfcp.

Bugzilla
--------
BZ 633525
https://bugzilla.redhat.com/show_bug.cgi?id=633525

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.37
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=d36deae75011a7890f0e730dd0f867c64081cb50

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>