qeth: NAPI support for l2 and l3 discipline
Patchwork-id: 29148
O-Subject: [RHEL6.1 PATCH 07/10] [s390x] qeth: NAPI support for l2 and l3
	discipline
Bugzilla: 633525

Description
-----------
This patch adds NAPI support to the qeth layer 2 and layer 3
discipline.  It is important to understand that it cannot
enable/disable IRQs as usual, we have to use the corresponding
new QDIO interface.  Also to not overdraw the budget we have to
stop and restart buffer processing at any point during processing
a bulk of QDIO buffers.  Having the driver NAPI enabled it is possible
to turn on GRO for the layer 3 discipline.

Bugzilla
--------
BZ 633525
https://bugzilla.redhat.com/show_bug.cgi?id=633525

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.37
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=a1c3ed4c9ca01dded8d511a1d1daf271fbae8d89

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>