Revalidate caches on lock
Patchwork-id: 29306
O-Subject: [RHEL6.1 PATCH 65/70] Revalidate caches on lock
Bugzilla: 653066
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Ricardo Labiaga <Ricardo.Labiaga@netapp.com>

Instead of blindly zapping the caches, attempt to revalidate them if
the server has indicated that it uses high resolution timestamps.

NFSv4 should be able to always revalidate the cache since the
protocol requires the update of the change attribute on modification of
the data.  In reality, there are servers (the Linux NFS server
for example) that do not obey this requirement and use ctime as the
basis for change attribute.  Long term, the server needs to be fixed.
At this time, and to be on the safe side, continue zapping caches if
the server indicates that it does not have a high resolution timestamp.

Signed-off-by: Ricardo Labiaga <Ricardo.Labiaga@netapp.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>