NFS: change stateid to be a union
Patchwork-id: 29314
O-Subject: [RHEL6.1 PATCH 04/18] NFS: change stateid to be a union
Bugzilla: 479351
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Alexandros Batsakis <batsakis@netapp.com>

In NFSv4.1 the stateid consists of the other and seqid fields. For layout
processing we need to numerically compare the seqid value of layout stateids.
To do so, introduce a union to nfs4_stateid to switch between opaque(16 bytes)
and opaque(12 bytes) / __be32

Signed-off-by: Alexandros Batsakis <batsakis@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
Signed-off-by: Fred Isaman <iisaman@netapp.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>