NFS: client needs to maintain list of inodes with active layouts
Patchwork-id: 29315
O-Subject: [RHEL6.1 PATCH 10/18] NFS: client needs to maintain list of inodes
	with active layouts
Bugzilla: 479351
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Andy Adamson <andros@netapp.com>

In particular, server reboot will invalidate all layouts.

Note that in order to have an active layout, we must get a successful response
from the server.  To avoid adding that machinery, this patch just includes a
stub that fakes up a successful return.  Since the layout is never referenced
for io, this is not a problem.

Signed-off-by: Andy Adamson <andros@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
Signed-off-by: Dean Hildebrand <dhildebz@umich.edu>
Signed-off-by: Fred Isaman <iisaman@netapp.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>