sunrpc: extract some common sunrpc_cache code from nfsd
Patchwork-id: 29338
O-Subject: [RHEL6.1 PATCH 01/96] sunrpc: extract some common sunrpc_cache code
	from nfsd
Bugzilla: 653068
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: NeilBrown <neilb@suse.de>

Rather can duplicating this idiom twice, put it in an inline function.
This reduces the usage of 'expiry_time' out side the sunrpc/cache.c
code and thus the impact of a change that is about to be made to that
field.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>