nfsd4: fix hang on fast-booting nfs servers
Patchwork-id: 29349
O-Subject: [RHEL6.1 PATCH 06/96] nfsd4: fix hang on fast-booting nfs servers
Bugzilla: 653068
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: J. Bruce Fields <bfields@redhat.com>

The last_close field of a cache_detail is initialized to zero, so the
condition

	detail->last_close < seconds_since_boot() - 30

may be false even for a cache that was never opened.

However, we want to immediately fail upcalls to caches that were never
opened: in the case of the auth_unix_gid cache, especially, which may
never be opened by mountd (if the --manage-gids option is not set), we
want to fail the upcall immediately.  Otherwise client requests will be
dropped unnecessarily on reboot.

Also document these conditions.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>