sunrpc: close connection when a request is irretrievably lost
Patchwork-id: 29344
O-Subject: [RHEL6.1 PATCH 07/96] sunrpc: close connection when a request is
	irretrievably lost.
Bugzilla: 653068
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: NeilBrown <neilb@suse.de>

If we drop a request in the sunrpc layer, either due kmalloc failure,
or due to a cache miss when we could not queue the request for later
replay, then close the connection to encourage the client to retry sooner.

Note that if the drop happens in the NFS layer, NFSERR_JUKEBOX
(aka NFS4ERR_DELAY) is returned to guide the client concerning
replay.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>