nfsd4: Move callback setup to callback queue
Patchwork-id: 29387
O-Subject: [RHEL6.1 PATCH 43/96] nfsd4: Move callback setup to callback queue
Bugzilla: 653068
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: J. Bruce Fields <bfields@citi.umich.edu>

Instead of creating the new rpc client from a regular server thread,
set a flag, kick off a null call, and allow the null call to do the work
of setting up the client on the callback workqueue.

Use a spinlock to ensure the callback work gets a consistent view of the
callback parameters.

This allows, for example, changing the callback from contexts where
sleeping is not allowed.  I hope it will also keep the locking simple as
we add more session and trunking features, by serializing most of the
callback-specific work.

This also closes a small race where the the new cb_ident could be used
with an old connection (or vice-versa).

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>