nfsd4: move callback setup into session init code
Patchwork-id: 29421
O-Subject: [RHEL6.1 PATCH 79/96] nfsd4: move callback setup into session init
	code
Bugzilla: 653068
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: J. Bruce Fields <bfields@citi.umich.edu>

The backchannel should  be associated with a session, it isn't really
global to the client.

We do, however, want a pointer global to the client which tracks which
session we're currently using for client-based callbacks.

This is a first step in that direction; for now, just reshuffling of
code with no significant change in behavior.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>