configs: enable CONFIG_PRINT_QUOTA_WARNING
Patchwork-id: 4506
O-Subject: [kernel team] [RHEL6 PATCH] configs: enable
	CONFIG_PRINT_QUOTA_WARNING
Bugzilla: 579638
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=579638

On RHEL5 when the user soft limit is exceeded, the kernel will print a
warning and this seems to be important to some customers. The warning is
controlled by a config variable, CONFIG_PRINT_QUOTA_WARNING, which is
disabled in RHEL6. This patch enables it.

Tested with rh-configs-prep.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>