cifs: map NT_STATUS_ERROR_WRITE_PROTECTED to -EROFS
Patchwork-id: 29554
O-Subject: [RHEL6.1 PATCH] BZ#618175: cifs: map NT_STATUS_ERROR_WRITE_PROTECTED
	to -EROFS
Bugzilla: 618175

When someone exports a CIFS share r/o, the server will generally send
error 19 (ERRwriteprot) in response to attempts to write to it. The
CIFS code in RHEL 6 currently maps this to -EIO, but -EROFS seems
like a more sensible mapping.

FWIW, the main impetus to this change was some odd behavior by older
"mv" programs, which interpreted EIO errors to mean that you were
trying to move a directory to be a subdirectory of itself. So
attempting to move a file on a r/o exported share would get you an
error like this:

    mv: cannot move `new' to a subdirectory of itself, `orig'

Newer "mv" programs don't seem to do this, but this error mapping
still makes more sense than EIO. This patch went into mainline in
2.6.36 in commit 3572d285.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>