gfs: Use 512 B block sizes to communicate with userland quota tools
Patchwork-id: 29690
O-Subject: [RHEL6.1 PATCH][GFS2] - Use 512 B block sizes to communicate with
	userland quota tools
Bugzilla: 658590
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The generic quota tools that communicate with gfs2 through the quotactl() syscall assume usage/limit/warn values to be in 512B block sizes instead of the fs block size that is assumed by gfs2.
This causes incorrect numbers to be reported and assigned. This patch fixes the problem.

This patch is also upstream.

Resolves: rhbz#658590
Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>