Disable paranoid check in ACPI power meter driver
Patchwork-id: 29720
O-Subject: [PATCH V2] [RHEL 6.1 PATCH] Disable paranoid check in ACPI power
	meter driver
Bugzilla: 634640
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

rhbz: #634640

The ACPI power meter driver supports power capping, but the driver currently
disables that and prints a concerning message unless the system is an IBM.
This appears to be undue paranoia caused by a misreading of the spec, so
lets just remove that check. Still discussing it with upstream.