GFS2: support for growing a full filesytem
Patchwork-id: 29790
O-Subject: [RHEL 6.1 PATCH] GFS2: support for growing a full filesytem (bz
	#659137)
Bugzilla: 659137
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

When GFS2 grew the filesystem, it was never rereading the rindex file during
the grow. This is necessary for large grows when the filesystem is almost full,
and GFS2 needs to use some of the space allocated earlier in the grow to
complete it.  Now, if GFS2 fails to reserve the necessary space and the rindex
data is not uptodate, it rereads it.  It also sets the rindex holder flags to
not cache the rindex glock, so that the rindex data is marked not uptodate when
the rindex file is grown.  Also, the only difference between gfs2_ri_update()
and gfs2_ri_update_special() was that gfs2_ri_update_special() didn't clear out
the existing resource groups, since you knew that it was only called when there
were no resource groups.  Attempting to clear out the resource groups when
there are none takes almost no time, and rarely happens, so there's no need for
gfs2_ri_update_special().

If the grow fails, GFS2 needs to trucate the rindex file to remove partial
rindex entries. When you truncate the rindex file, you need to avoid calling
gfs2_rindex_hold, since you already hold it.  However, if you haven't already
read in the resource groups, you need to do that.

This fixes bz #659137
This is already in Steve's upstream git repository.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>