cifs: eliminate redundant xdev check in cifs_rename
Patchwork-id: 29851
O-Subject: [RHEL6.1 PATCH 02/27] BZ#646223: cifs: eliminate redundant xdev check
	in cifs_rename
Bugzilla: 646223

The VFS always checks that the source and target of a rename are on the
same vfsmount, and hence have the same superblock. So, this check is
redundant. Remove it and simplify the error handling.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 639e7a913d81f918bfbf506e6ecd54664f787cbd

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>