cifs: add cifs_sb_master_tcon and convert some callers to use it
Patchwork-id: 29854
O-Subject: [RHEL6.1 PATCH 08/27] BZ#646223: cifs: add cifs_sb_master_tcon and
	convert some callers to use it
Bugzilla: 646223

At mount time, we'll always need to create a tcon that will serve as a
template for others that are associated with the mount. This tcon is
known as the "master" tcon.

In some cases, we'll need to use that tcon regardless of who's accessing
the mount. Add an accessor function for the master tcon and go ahead and
switch the appropriate places to use it.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 0d424ad0a4b8c08e45928bccfa5b4b240097b01b

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>