cifs: add refcounted and timestamped container for holding tcons
Patchwork-id: 29858
O-Subject: [RHEL6.1 PATCH 12/27] BZ#646223: cifs: add refcounted and timestamped
	container for holding tcons
Bugzilla: 646223

Eventually, we'll need to track the use of tcons on a per-sb basis, so that
we know when it's ok to tear them down. Begin this conversion by adding a
new "tcon_link" struct and accessors that get it. For now, the core data
structures are untouched -- cifs_sb still just points to a single tcon and
the pointers are just cast to deal with the accessor functions. A later
patch will flesh this out.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 7ffec372458d163492e56e663a1b3a2d7be0a0a2

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>