cifs: have cifsFileInfo hold a reference to a tlink rather than tcon pointer
Patchwork-id: 29869
O-Subject: [RHEL6.1 PATCH 13/27] BZ#646223: cifs: have cifsFileInfo hold a
	reference to a tlink rather than tcon pointer
Bugzilla: 646223

cifsFileInfo needs a pointer to a tcon, but it doesn't currently hold a
reference to it. Change it to keep a pointer to a tcon_link instead and
hold a reference to it.

That will keep the tcon from being freed until the file is closed.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 13cfb7334eb6fd0fc06da5589aea1e947791f1d6

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>