cifs: store pointer to master tlink in superblock
Patchwork-id: 29868
O-Subject: [RHEL6.1 PATCH 23/27] BZ#646223: cifs: store pointer to master tlink
	in superblock (try #2)
Bugzilla: 646223

This is the second version of this patch, the only difference between
it and the first one is that this explicitly makes cifs_sb_master_tlink
a static inline.

Instead of keeping a tag on the master tlink in the tree, just keep a
pointer to the master in the superblock. That eliminates the need for
using the radix tree to look up a tagged entry.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 413e661c136c52290de1ee19a1b049a4da9dbf51

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>