cifs: don't take extra tlink reference in initiate_cifs_search
Patchwork-id: 29872
O-Subject: [RHEL6.1 PATCH 27/27] BZ#646223: cifs: don't take extra tlink
	reference in initiate_cifs_search
Bugzilla: 646223

It's possible for initiate_cifs_search to be called on a filp that
already has private_data attached. If this happens, we'll end up
calling cifs_sb_tlink, taking an extra reference to the tlink and
attaching that to the cifsFileInfo. This leads to refcount leaks
that manifest as a "stuck" cifsd at umount time.

Fix this by only looking up the tlink for the cifsFile on the filp's
first pass through this function. When called on a filp that already
has cifsFileInfo associated with it, just use the tlink reference
that it already owns.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Upstream commit: 59c55ba1fba61ae13fbfff094a674e73407201f1

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>