kernel: fix clock comparator revalidation
Patchwork-id: 29909
O-Subject: [RHEL6.1 PATCH 03/14] [s390x] kernel: fix clock comparator
	revalidation
Bugzilla: 658578

Description
-----------
Kernel hangs for a short or long period of time if a machine check
occurs, for example, if a device gets attached or detached.

On each machine check all registers are revalidated. The save
area for the clock comparator however only contains the upper
most seven bytes of the former contents, if valid.
Therefore the machine check handler uses a store clock instruction
to get the current time and writes that to the clock comparator
register which in turn will generate an immediate timer interrupt.
However within the lowcore the expected time of the next timer
interrupt is stored. If the interrupt happens before that time the
handler won't be called. In turn the clock comparator won't be
reprogrammed and therefore the interrupt condition stays pending
which causes an interrupt loop until the expected time is reached.

To fix this just revalidate the clock comparator register with the
expected value or the current time if udelay handling is in
progress. In addition the special handling for udelay must be
changed as well.

Bugzilla
--------
BZ 658578
https://bugzilla.redhat.com/show_bug.cgi?id=658578

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.37
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=e8129c642155616d9e2160a75f103e127c8c3708

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>