allow setting of phys_device (memory hotplug)
Patchwork-id: 29910
O-Subject: [RHEL6.1 PATCH 04/14] [memory] kernel: allow setting of phys_device
	(memory hotplug)
Bugzilla: 659474

Description
-----------
/sys/devices/system/memory/memoryX/phys_device is supposed to
contain the number of the physical device that the corresponding
piece of memory belongs to.

If a physical device should be replaced or taken offline for whatever
reason it is necessary to set all corresponding memory pieces offline.
The current implementation always sets phys_device to '0' and there is
no way or hook to change that.

Add a weak function which architectures can override to actually set
the phys_device from within add_memory_block().

Bugzilla
--------
BZ 659474
https://bugzilla.redhat.com/show_bug.cgi?id=659474

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.34
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=bc32df00894f0e1dbf583cc3dab210d2969b078a

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>