dasd: improve error recovery for internal I/O
Patchwork-id: 29911
O-Subject: [RHEL6.1 PATCH 07/14] [s390x] dasd: improve error recovery for
	internal I/O
Bugzilla: 631497

Description
-----------
Most of the error conditions reported by a FICON storage server
indicate situations which can be recovered. Sometimes the host
just needs to retry an I/O request, but sometimes the recovery
is more complex and requires the device driver to wait, choose
a different path, etc.

The DASD device driver has a fully featured error recovery
for normal block layer I/O, but not for internal I/O request
which are for example used during the device bring up.
This can lead to situations where the IPL of a system fails
because DASD devices are not properly recognized.
This patch will extend the internal I/O handling to use the
existing error recovery procedures.

Bugzilla
--------
BZ 631497
https://bugzilla.redhat.com/show_bug.cgi?id=631497

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=eb6e199bef288611157b8198c25d12b32bf058d0

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>